// src/types/generated/RenderScene.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Animal } from "./Animal";
import type { AnimalRegionLayout } from "./AnimalRegionLayout";
import type { SpeciesProfile } from "./SpeciesProfile";

export type RenderScene = { animals: Array<Animal>, regions: Array<AnimalRegionLayout>, speciesProfiles: Array<SpeciesProfile>, };